% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate the shortest distance between two points on an ellipsoid}
\usage{
calculate_distance(x_lon, x_lat, y_lon, y_lat)
}
\arguments{
\item{x_lon}{\itemize{
\item x longitude
}}

\item{x_lat}{\itemize{
\item x latitude
}}

\item{y_lon}{\itemize{
\item y longitude
}}

\item{y_lat}{\itemize{
\item y latitude
}}
}
\value{
distance between 2 points
}
\description{
Calculate the shortest distance between two points on an ellipsoid
}
