% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_preprocess.R
\name{general_data_preprocess}
\alias{general_data_preprocess}
\title{General Data preprocessing function}
\usage{
general_data_preprocess(df1)
}
\arguments{
\item{df1}{\itemize{
\item dataframe with ship data
}}
}
\value{
data_list - list with multiple parameters
ranks, next latitude, next longitude
}
\description{
General Data preprocessing function
}
